move format function
